+++ test/integration/kernel_test.rb
@@ -8,6 +8,8 @@ module Bootsnap
     include TmpdirHelper
 
     def test_require_symlinked_file_twice
+      skip("https://github.com/oracle/truffleruby/issues/3138") if truffleruby?
+
       setup_symlinked_files
       if RUBY_VERSION >= "3.1"
         # Fixed in https://github.com/ruby/ruby/commit/79a4484a072e9769b603e7b4fbdb15b1d7eccb15 (Ruby 3.1)
@@ -41,6 +43,8 @@ module Bootsnap
     end
 
     def test_require_relative_symlinked_file_twice
+      skip("https://github.com/oracle/truffleruby/issues/3138") if truffleruby?
+
       setup_symlinked_files
       if RUBY_VERSION >= "3.1"
         # Fixed in https://github.com/ruby/ruby/commit/79a4484a072e9769b603e7b4fbdb15b1d7eccb15 (Ruby 3.1)
@@ -83,6 +87,8 @@ module Bootsnap
     end
 
     def test_require_deep_symlinked_file_twice
+      skip("https://github.com/oracle/truffleruby/issues/3138") if truffleruby?
+
       setup_symlinked_files
       if RUBY_VERSION >= "3.1"
         # Fixed in https://github.com/ruby/ruby/commit/79a4484a072e9769b603e7b4fbdb15b1d7eccb15 (Ruby 3.1)
@@ -116,6 +122,8 @@ module Bootsnap
     end
 
     def test_require_relative_deep_symlinked_file_twice
+      skip("https://github.com/oracle/truffleruby/issues/3138") if truffleruby?
+
       setup_symlinked_files
       if RUBY_VERSION >= "3.1"
         # Fixed in https://github.com/ruby/ruby/commit/79a4484a072e9769b603e7b4fbdb15b1d7eccb15 (Ruby 3.1)
@@ -210,6 +218,10 @@ module Bootsnap
         end
       RUBY
       File.symlink("real", "symlink")
+    end
+
+    def truffleruby?
+      RUBY_ENGINE == "truffleruby"
     end
   end
 end